Vastly improved file upload
Files are now uploaded under a uuid name that is made available in the
params to the response function. File writing now respects body-offset
when writing buffers, so files are written correctly. Both large and
small files are supported and the difference is transparent to the
response function. Will need to wait until yada 1.1.32 to get good
performance on large files.