Adding header to csv upload
As part of this file metadata is now conformed on upload. Which required
changing the spec for provenance from an or to a multispec. With or the
conformer returns a vector of the valid spec and the value, which
effectively spoils the data. As a multi it keeps it the same.

I also added the notion of a boolean to our conformers.

A bit of a performance tweak to the ordering of the integer? conformer
happened as well, probably a bit more fat in there yet though.