Generative tests modified to be more representative of the style
Added test.chuck which provies a checking macro, this allows writing of
generative tests that use 'is'. The tests now correctly report their
number and, when failing, shrink the input and report the problem
properly.

Removed the requirement for kafka to be running, the tests now just
directly use the schemastore and don't go via the event stream.