Sharing metadata definded and working into file post
The file and file-metadata permissions have been defined and added to
the spec's for the file-metadata.

A problem with our uuid spec type has been solved, lack of a call to
conformer to wrap the base definition was causing valid? to see
:clojure.spec/invalid which it interrpretted as true.

Finally solved the problem with metrics having a non-typed call, which
was throwing off the compilation