File sharing
File permission and associated auth code has been added to the
components.

The Sharing component aims to be a central place for all the auth
routing that's going to be required and a good place to centralise
metrics. It looks a bit over kill right now, but figure once Schema
permissions come it, it will start to pay off a bit.

All the tests had to be slightly modified to make them submit id's to
access all the things they were creating.