Breaking cyclic nature of the list spec definition
The list spec allows for lists to contain lists, which seems to be
driving the generators a bit crazy. Here that cyclic link has been
removed.

Moved the schema transport to internal format code into the transport
namespace.

Tidied up more of the testing functions to try and make them all
similar, not quite there yet. Testing 'when' macro pprints the output
when the status is not as expected. Should move that to the 'is' call really.

This is a breaking commit as nothing is quite working again yet.