Implement S3 storage for filestore
There is no reliable local S3 implementation at this time, due to the
java sdk being up graded recently to move to chunked upload.

Therefore, when testing locally, by default the s3 layer will not be
used. Need to find a way to get this work correctly on circleci at
least.

There is a bit to be desired around the code at this point. There is no
S3 upload error handling, and the use of a tuple in the return from
output-stream is not very good. Might have to introduce more core.async
to make that clean or just live with this until a better idea occurs.