Refine handlers for metadata and schema creation in tests
Previously the test handlers just waited for the first message they
received and treated that like the response. This worked fine until a
new test didn't use the handlers, creating a backlog for susequent
tests.

Now the test handlers look for events to do with the specific entity
they have just uploaded. This is based around the id of the uploading
user, some tests reuse this id, so it' not water tight, but if the fix
will be simple in the future.