Local and jenkins testing moved to inmemory comms (#82)
A cheap win to make most of our testing much faster. When the tests
are run both locally and on Jenkins the system will use inmemory
coreasync communication.

When the service is deployed all the tests will then go over
Kinesis. To make that bit faster we could run a subset of the tests.