Creation and Edit of Datapacks
To create a datapack you must have meta-read on all packed files.

All updating is done on a replace basis, like files. There are some
caveats around updating the packed file ids. The cmd is only allowed
to add or remove file-ids that the user can read. This means that when
updating the user must send all the file ids that they can current
read, with the new additions. To remove file's they just ommit the
relavent id's.