Configuring web_server and local file storage
Need the webserver for the marathon health checks.

The local file store is a bit of a cheat, easier to just have one
rather than modify the code to allow for it to be missing. Local so
that it doesn't really work as in replay mode datastore shouldn't be
able to work with actual files.