System components now optional
Rather than work around in the comms library or in the components the
fact that the command channals might not be there when in replay. This
changes makes it so that service components can be unconfigured.

As our Command Handlers should be there own components (schema is not
yet), this allows the replay mode config to just ignore them.

Components that require, via the dependancy map, a missing component
just don't get it. Obviously this will led to NPE's if attempts are
made to use them. Ideally, we'd solve this by having breaking down the
compoennts some more, but this feels like far enough for now.