The purpose of the consolidation is to avoid having two separate processes for upload.
Previously, once the user had established whether their file is greater-than the minimum size for a multipart upload they would use two entirely separate exchanges. Now there is just one; the juggling of mpu-or-not is handled entirely by the Datastore. This costs us a bit of state which is now handled by DynamoDB.