Adding cheshire encodes so exceptions are treated properly
Added a test to inject a through put exception from dynamo. This made
it clear that exceptions were going to be logged three times, so
removed the custom yada exception logger. The exceptions still get
logged twice, but the first time is quite light, the second has all
the details we want in it.