Moving mup cache to command handler
Currently the multi part upload cache is written to when an initate
upload event is consumed by an event handler. This is occuring too late
for quick uploads and they are being failed due to the line not having
been written yet.

This change writes to the mup cache when the command is deemed
successful, therefore the line will already be present before the
upload can possibly be completed.