Rejected events use 'relaxed' versions of specs
In the instance that a command is received and deemed to be invalid,
we cannot simply pass data through; it should be assumed that all data
is garbage, unparse-able etc. In this case, it only makes sense to use
less strict spec when we send out 'rejected' events that typically
reference this data.