Add new format commands/events
As much as possible, the old code paths have been adapted, although very little is needed as the payload of the original commands/events is the spec of the new commands/events. Tests have also been added to ensure the old code paths and old commands/events are still working. Until the old commands/events are truly deprecated, there might be parts of the system still using them.